reconfigured to use sessions. now prevents users from reaccess.
